compiler issue fixed: add class name before function
