+++ Database Project 1/include/BplusTree.h
@@ -38,7 +38,8 @@ struct BTreeNode{
 class Btree
 {
 
-    private:
+    public:
+    // private:
         int deg; //degree
         BTreeNode* root;
         int nodeCount;
@@ -47,25 +48,29 @@ class Btree
         void insertParent(Key_Records key, BTreeNode *current, BTreeNode *child);
         BTreeNode *findParent( BTreeNode *current, BTreeNode *child);
 
-
-
         void removeRecord(int key, BTreeNode* node);
+
+        void remove_key_in_leaf_node(BTreeNode* leafNode, int key);
+        void insert_in_leaf_node(BTreeNode* leafNode, int key, vector<RecordAddress>* addressVector);
+
 
         bool tryBorrowing(BTreeNode* node1, BTreeNode* node2);
 
         int mergeTwoNodes(BTreeNode* node1, BTreeNode* node2);
 
-        void insert_in_leaf_node(BTreeNode* leafNode, int key, vector<RecordAddress>* addressVector);
 
-        void remove_key_in_leaf_node(BTreeNode* leafNode, int key);
 
         void remove_key_in_internal_node(BTreeNode* internalNode, int key);
 
         void insert_ChildNode_in_ParentNode(BTreeNode* parent, BTreeNode* child, int key);
+        
+        void findParentAndIndex(BTreeNode* root, BTreeNode* current, BTreeNode* prev, BTreeNode* nodeToFind, BTreeNode*& parent, int& index);
+
+        std::pair<BTreeNode*, BTreeNode*> findAdjacentSiblings(BTreeNode* root, BTreeNode* node);
 
 
         
-    public:
+    // public:
         Btree(unsigned int BlockSize){
             root = nullptr;
             nodeCount =0;
